requirements.txt
hydra-core==1.3.2
numpy==1.26.4
pandas==2.2.1
torch==2.2.1
stable-baselines3==2.2.1
imageio==2.34.0
black==24.3.0
notebook==7.1.2

swig==4.2.1
gymnasium==0.29.1